api-python/requirements.txt
fastapi
uvicorn[standard]
gunicorn
groq
openai
anthropic
python-dotenv
pydantic
httpx
chromadb
sentence-transformers
numpy
